requirements.txt
playwright
beautifulsoup4
markdownify
fastapi
uvicorn
pydantic
transformers
pytesseract
Pillow